+++ packages/core/src/colors.ts
@@ -171,12 +171,18 @@ function getLocationAreaColorsRGBA(colors: LocationAreaColors | undefined, darkM
   return {
     normal: locationAreasNormal,
     connected: colorAsRgbaOr(colors && colors.connected, locationAreasNormal),
-    highlighted: colorAsRgbaOr(colors && colors.highlighted, locationAreasNormal),
+    highlighted: colorAsRgbaOr(
+      colors && colors.highlighted,
+      opacifyHex(normalColorHcl[darkMode ? 'brighter' : 'darker'](1).toString(), 0.5),
+    ),
+    selected: colorAsRgbaOr(
+      colors && colors.selected,
+      opacifyHex(normalColorHcl[darkMode ? 'brighter' : 'darker'](2).toString(), 0.8),
+    ),
     outline: colorAsRgbaOr(
       colors && colors.outline,
-      colorAsRgba(normalColorHcl[darkMode ? 'brighter' : 'darker']().toString()),
+      colorAsRgba(normalColorHcl[darkMode ? 'brighter' : 'darker'](4).toString()),
     ),
-    selected: colorAsRgbaOr(colors && colors.selected, locationAreasNormal),
   };
 }
 
@@ -240,6 +246,16 @@ function colorAsRgbaOr(color: string | undefined, defaultColor: RGBA | string): 
 
 export function rgbaAsString(color: RGBA): string {
   return `rgba(${color.join(',')})`;
+}
+
+export function opacifyHex(hexCode: string, opacity: number): string {
+  const c = d3color(hexCode);
+  if (!c) {
+    console.warn('Invalid color: ', hexCode);
+    return `rgba(255, 255, 255, ${opacity})`;
+  }
+  const col = c.rgb();
+  return `rgba(${col.r}, ${col.g}, ${col.b}, ${opacity})`;
 }
 
 export function opacityFloatToInteger(opacity: number): number {